backend/requirements.txt
Django==3.2.3
django-colorfield==0.10.1
django-import-export==3.3.1
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
webcolors==1.11.1
